connection with the database
