+++ src/main/java/com/example/productrepository/Product/Controller/ProductController.java
@@ -1,0 +1,27 @@
+package com.example.productrepository.Product.Controller;
+
+import com.example.productrepository.Product.Entities.Product;
+import com.example.productrepository.Product.Services.ProductService;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("api/products")
+public class ProductController {
+    private final ProductService productService;
+
+    public ProductController(ProductService productService) {
+        this.productService = productService;
+    }
+
+    @GetMapping()
+    public List<Product> getAllProducts(){
+        return productService.getAllProducts();
+    }
+
+
+}
