+++ src/main/java/com/example/productrepository/Product/Entities/Product.java
@@ -1,4 +1,4 @@
-package com.example.productrepository;
+package com.example.productrepository.Product.Entities;
 
 public record Product(String id, String title, int price) {
 }
