+++ src/main/java/com/example/productrepository/Product/Repository/ProductRepository.java
@@ -1,5 +1,6 @@
-package com.example.productrepository;
+package com.example.productrepository.Product.Repository;
 
+import com.example.productrepository.Product.Entities.Product;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface ProductRepository extends MongoRepository<Product, String> {
