--- src/main/java/com/example/productrepository/ProductController.java
@@ -1,23 +1,0 @@
-package com.example.productrepository;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("api/products")
-public class ProductController {
-    private final ProductRepository productRepository;
-
-    public ProductController(ProductRepository productRepository) {
-        this.productRepository = productRepository;
-    }
-
-    @GetMapping()
-    public List<Product> getAllProducts(){
-      return this.productRepository.findAll();
-    }
-
-}
