+++ src/main/resources/application.properties
@@ -1,1 +1,2 @@
 spring.data.mongodb.uri=${MONGO_DB_URI}
+spring.data.mongodb.database=shop
